Change monitoring paramters time to uint64

+++ Framework/include/QualityControl/BenchmarkDevices.h
@@ -24,21 +24,21 @@ class GeneratorDevice : public framework::Task {
     framework::Outputs getOutputs() { return { mOutputSpec }; }
     GeneratorDevice(const GeneratorDevice& device):GeneratorDevice(device.mName, device.mConfigSource){}
   private:
-    const int second = 1000000;
+    const long second = 1000000;
 
     std::string mName;
     std::string mConfigSource;
     int mSize = 100; // number of elements
     int mFrequency = 10; // usec
     int mDuration = 100; // seconds
-    int mQuantity = 0;
-    int mLastQuantity = 0;
+    uint64_t mQuantity = 0;
+    uint64_t mLastQuantity = 0;
     int mDelay = 30; // seconds
     TArrayD* mObject;
     bool mFirstTime = true;
     int mAdaptivePeriod = 300; 
 
-    const int mStatPeriod = 10 * second;
+    const long mStatPeriod = 10 * second;
     
     AliceO2::Common::Timer mExperimentTimer;
     AliceO2::Common::Timer mStatTimer;
@@ -61,13 +61,13 @@ class CollectorDevice : public framework::Task {
   private:
     framework::Inputs mInputs;
     std::map<std::string, int> mCollected;
-    int mGlobalReceived = 0;
+    uint64_t mGlobalReceived = 0;
     std::unique_ptr<o2::configuration::ConfigurationInterface> mConfigFile; // used in init only
     std::unique_ptr<o2::monitoring::Monitoring> mMonitoring;
     AliceO2::Common::Timer mStatTimer;
 
 
-    const int mStatPeriod = 10 * 1000 * 1000;
+    const long mStatPeriod = 10 * 1000 * 1000;
 };
 
 }
